# storage/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Oja Ewa Pro",
    "version": "1.0.0",
    "description": "Comprehensive e-commerce API for the Oja Ewa Pro platform",
    "contact": {
      "name": "API Support",
      "email": "support@ojaewa.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local Development Server"
    },
    {
      "url": "https://test.ojaewa.com/api",
      "description": "Production Server"
    }
  ],
  "paths": {
...
    "/profile": {
      "get": {
        "tags": ["User Profile"],
        "summary": "Get user profile",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "Accept", "in": "header", "required": true, "schema": {"type": "string", "default": "application/json"}}],
        "responses": {"200": {}}
      },
      "put": {
        "tags": ["User Profile"],
        "summary": "Update user profile",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "Accept", "in": "header", "required": true, "schema": {"type": "string", "default": "application/json"}}],
        "responses": {"200": {}}
      }
    },
...
    "/orders/{order}": {
      "get": {
        "tags": ["User Orders"],
        "summary": "Get order",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "order", "in": "path", "required": true, "schema": {"type": "integer"}},
          {"name": "Accept", "in": "header", "required": true, "schema": {"type": "string", "default": "application/json"}}
        ],
        "responses": {"200": {}}
      }
    },
...
    "/seller/profile": {
      "get": {
        "tags": ["User Seller"],
        "summary": "Get seller profile",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "Accept", "in": "header", "required": true, "schema": {"type": "string", "default": "application/json"}}],
        "responses": {"200": {}}
      },
      "post": {
        "tags": ["User Seller"],
        "summary": "Create seller profile",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "Accept", "in": "header", "required": true, "schema": {"type": "string", "default": "application/json"}}],
        "responses": {"201": {}}
      },
      "put": {
        "tags": ["User Seller"],
        "summary": "Update seller profile",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "Accept", "in": "header", "required": true, "schema": {"type": "string", "default": "application/json"}}],
        "responses": {"200": {}}
      },
      "delete": {
        "tags": ["User Seller"],
        "summary": "Delete seller profile",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "Accept", "in": "header", "required": true, "schema": {"type": "string", "default": "application/json"}}],
        "responses": {"200": {}}
      }
    },
...
    "/admin/profile": {
      "get": {
        "tags": ["Admin Auth"],
        "summary": "Get admin profile",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "Accept", "in": "header", "required": true, "schema": {"type": "string", "default": "application/json"}}],
        "responses": {"200": {}}
      }
    },
...
    "/admin/market/sellers": {
      "get": {
        "tags": ["Admin Market"],
        "summary": "List sellers",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "Accept", "in": "header", "required": true, "schema": {"type": "string", "default": "application/json"}}],
        "responses": {"200": {}}
      }
    },
    "/admin/pending/sellers": {
      "get": {
        "tags": ["Admin Market"],
        "summary": "List pending sellers",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "Accept", "in": "header", "required": true, "schema": {"type": "string", "default": "application/json"}}],
        "responses": {"200": {}}
      }
    },
    "/admin/sellers/{id}": {
      "get": {
        "tags": ["Admin Market"],
        "summary": "Get seller details",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}},
          {"name": "Accept", "in": "header", "required": true, "schema": {"type": "string", "default": "application/json"}}
        ],
        "responses": {"200": {}}
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}